src/third_party/blink/web_tests/external/wpt/css/css-scroll-snap/snap-inline-block-expected.txt

This is a testharness.js-based test.
PASS Snaps correctly for horizontal-tb writing mode with 'inline' and 'block' alignments
PASS Snaps correctly for vertical-lr writing mode with 'inline' and 'block' alignments
FAIL Snaps correctly for vertical-rl writing mode with 'inline' and 'block' alignments assert_equals: aligns correctly on x expected -315 but got 300
Harness: the test ran to completion.
